voice-clone-tts/requirements-nemo-msdd.txt
# NeMo MSDD Speaker Diarization Dependencies
# 安装说明见: docs/NEMO-MSDD-SETUP.md

# Core NeMo toolkit with ASR support
nemo_toolkit[asr]>=1.23.0

# Audio processing
torchaudio>=2.0.0
soundfile>=0.12.1
librosa>=0.10.0

# NeMo dependencies (if not auto-installed)
omegaconf>=2.3.0
hydra-core>=1.3.0
pytorch-lightning>=2.0.0

# Optional: For better performance
numba>=0.57.0
